+++ src/BlockArea.cpp
@@ -738,31 +738,31 @@ void cBlockArea::Fill(int a_DataTypes, BLOCKTYPE a_BlockType, NIBBLETYPE a_Block
 		a_DataTypes = a_DataTypes & GetDataTypes();
 	}
 	
-	int BlockCount = GetBlockCount();
+	size_t BlockCount = GetBlockCount();
 	if ((a_DataTypes & baTypes) != 0)
 	{
-		for (int i = 0; i < BlockCount; i++)
+		for (size_t i = 0; i < BlockCount; i++)
 		{
 			m_BlockTypes[i] = a_BlockType;
 		}
 	}
 	if ((a_DataTypes & baMetas) != 0)
 	{
-		for (int i = 0; i < BlockCount; i++)
+		for (size_t i = 0; i < BlockCount; i++)
 		{
 			m_BlockMetas[i] = a_BlockMeta;
 		}
 	}
 	if ((a_DataTypes & baLight) != 0)
 	{
-		for (int i = 0; i < BlockCount; i++)
+		for (size_t i = 0; i < BlockCount; i++)
 		{
 			m_BlockLight[i] = a_BlockLight;
 		}
 	}
 	if ((a_DataTypes & baSkyLight) != 0)
 	{
-		for (int i = 0; i < BlockCount; i++)
+		for (size_t i = 0; i < BlockCount; i++)
 		{
 			m_BlockSkyLight[i] = a_BlockSkyLight;
 		}
